This is a 50 character title, use it as a template Fix screen flashing when showing windows
When displaying the splash and the main windows for the first time,
don't show them until they are fully rendered to avoid displaying an
empty background.
Electron has a 'ready-to-show' event, but it seems it is dispatched a
little premature, thus a delay was added before dispatching the final
'loaded' event.